update: change folder structure for dev
